third_party/blink/web_tests/wpt_internal/fenced_frame/input-on-coop-fenced-frame.https.html

<!DOCTYPE html>
<title>Test input inside a fenced frame with non-matching COOP.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/utils.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<body>
<script>

// Navigating a fenced frame to a page with a COOP mismatch (in this case,
// unsafe-none -> same-origin-allow-popups) will force a browsing instance swap
// which means we'll force using a new RenderFrameHost. This will create a
// speculative frame host which will create the renderer's frame using
// RenderFrameImpl::CreateFrame rather than RenderFrameImpl::CreateMainFrame.
// This test is a basic smoke test that the WebFrameWidgetImpl is correctly
// setup in that path. This test sends some input in to exercise this code
// path, mainly ensuring the renderer doesn't crash/DCHECK.  See
// https://crgbug.com/1316535.
promise_test(async () => {
  const frame = attachFencedFrameContext({
    headers: [["Cross-Origin-Opener-Policy", "same-origin-allow-popups"]]
  });

  await frame.execute(async () => {
    if (document.readyState !== 'complete') {
      await new Promise((resolve) => { addEventListener('load', resolve); });
    }

    assert_equals(document.readyState, 'complete', 'Fenced frame was loaded');

    window.was_clicked = false;
    addEventListener('click', () => {
      window.was_clicked = true;
    });
  });

  await test_driver.click(frame.element);

  // Ensure the fenced frame actually saw the click but the real test is that
  // there isn't a crash.
  await frame.execute(async () => {
    assert_equals(window.was_clicked, true, 'Fenced frame received click');
  });

}, 'Input in non-matching COOP fenced frame doesn\'t crash.');
</script>
</body>
